requirements.txt
fastapi>=0.110
uvicorn[standard]
pydantic-settings>=2.0
crewai>=0.30
gspread>=6.0
google-auth>=2.0
requests>=2.31
apscheduler>=3.10
python-dotenv
python-jobspy>=1.1
